requirements/test.txt
django-nose>=1.3,<1.4
coverage>=3.7,<3.8
factory-boy>=2.4,<2.5
tox>=1.8,<1.9
